Refactor code structure for boss fights
