+++ bosses.py
@@ -1,0 +1,37 @@
+import deck
+import duel
+import QoL
+import card_library
+import ASCII_text
+import menu
+import card
+import random
+import sigils
+import os
+import rogue
+
+##### bosses will use the basic AI, but with higher difficulty settings and have their unique mechanics (pickaxe, ship, extra sigils, moon)
+
+def boss_fight_prospector(campaign) : # boss fight 1
+    def gameplay() :
+        rogue.card_battle() # for testing prior to implementation
+
+    gameplay() # add flavor text, context, etc.
+
+def boss_fight_angler(campaign) : # boss fight 2
+    def gameplay() :
+        rogue.card_battle() # for testing prior to implementation
+
+    gameplay() # add flavor text, context, etc.
+
+def boss_fight_trapper_trader(campaign) : # boss fight 3
+    def gameplay() :
+        rogue.card_battle() # for testing prior to implementation
+
+    gameplay() # add flavor text, context, etc.
+
+def boss_fight_leshy(campaign) : # boss fight 4
+    def gameplay() :
+        rogue.card_battle() # for testing prior to implementation
+        
+    gameplay() # add flavor text, context, etc.
